server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.121.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
